[Any ideas?] Add some test codes (but commented-out all). Devs! Please see below:
These work-in-progress test codes are for trying a new way to insert white-spaces of page-breaking inside documents whose directions of page-progression and line-progression are in the same, without destractive changings of the documents' DOM.
In the new way, shape-outside of CSS Shapes is used instead of column-gap of CSS Columns (cf. (-webkit-)column-axis has discontinued long ago).
But it remains many problems (ex. Chrome has bugs around laying-out of a child element which has different writing-mode from the parent).
If you have any ideas, work-arounds, or samples about it, please let me know!